wip: restructuring
