+++ packages/indexer/src/v3_indexer/builders/swaps.ts
@@ -26,7 +26,7 @@ import { logger } from "../../logger";
 import { getMainIxTypeFromTransaction } from "../transaction/watcher";
 import { getHumanPrice } from "../usecases/math";
 import { connection } from "../connection";
-import { AmmMarketAccountUpdateIndexer } from '../indexers/amm-market/amm-market-account-indexer';
+import { AmmMarketAccountUpdateIndexer } from '../indexers/amm/amm-market-account-indexer';
 import { PublicKey } from "@solana/web3.js";
