+++ packages/indexer/src/v3_indexer/indexers/start-interval-fetch-indexers.ts
@@ -7,7 +7,7 @@ import { IndexerWithAccountDeps } from "../types";
 import { BirdeyePricesIndexer } from "./birdeye/birdeye-prices-indexer";
 import { IntervalFetchIndexer } from "./interval-fetch-indexer";
 import { JupiterQuotesIndexer } from "./jupiter/jupiter-quotes-indexer";
-import { AmmMarketAccountIntervalFetchIndexer } from "./amm-market/amm-market-account-interval-indexer";
+import { AmmMarketAccountIntervalFetchIndexer } from "./amm/amm-market-account-interval-indexer";
 import { AutocratDaoIndexer } from "./autocrat/autocrat-dao-indexer";
 import { AutocratProposalIndexer } from "./autocrat/autocrat-proposal-indexer";
 import { TokenMintIndexer } from "./token/token-mint-indexer";
