+++ packages/indexer/src/v3_indexer/indexers/start-logs-subscribe-indexer.ts
@@ -3,7 +3,7 @@ import { PublicKey } from "@solana/web3.js";
 import { connection } from "../connection";
 import { IndexerWithAccountDeps } from "../types";
 import { AccountLogsIndexer } from "./account-logs-indexer";
-import { AmmMarketLogsSubscribeIndexer } from "./amm-market/amm-market-logs-subscribe-indexer";
+import { AmmMarketLogsSubscribeIndexer } from "./amm/amm-market-logs-subscribe-indexer";
 import { logger } from "../../logger";
 
 export async function startLogsSubscribeIndexer(
