+++ packages/indexer/src/v3_indexer/indexers/start-transaction-history-indexers.ts
@@ -6,7 +6,7 @@ import {
 } from "@metadaoproject/indexer-db/lib/schema";
 
 import { IndexerWithAccountDeps } from "../types";
-import { AmmMarketInstructionsIndexer } from "./amm-market/amm-market-instruction-indexer";
+import { AmmMarketInstructionsIndexer } from "./amm/amm-market-instruction-indexer";
 import { InstructionIndexer } from "./instruction-indexer";
 import { Idl } from "@coral-xyz/anchor";
 import { schema, usingDb, eq, and, gte } from "@metadaoproject/indexer-db";
